- Added patches from IBM which won't make it for 2.6.0;   enabled for S390 only for now. - Update s390 and s390x configuration.
suse-commit: 682b4f6ab2668e9add8b76e82666e7f190dfac07

+++ include/asm-s390/spinlock.h
@@ -217,7 +217,7 @@ typedef struct {
 
 extern inline int _raw_write_trylock(rwlock_t *rw)
 {
-	unsigned int result, reg;
+	unsigned long result, reg;
 	
 	__asm__ __volatile__(
 #ifndef __s390x__
